Major updates to project, with enhancements and bug fixes.
This commit is significant enough to warrant a name change. My changes have focused largely on making this project work with a mkdocs project. What motivated me to make these changes, was a lack of a program that could test all of the local and remote links found inside of mkdocs. Changes to the original forked repository include:

* Broke out the process of extracting links
* Enabled recursive file processing for both local and remote links
* Fixed apparent bug which caused files in child directories to be parsed when recurse=false
* Created a much more robust summary report at the end of the process
* Added the ability to exclude some files/urls from being checked
* Parsed not only markdown links, but HTML links as well